fix: Audit log static reference (#7703)
* fix: sort into new file

* refactor: move other stuff

* fix: errors
