+++ packages/discord.js/src/index.js
@@ -103,7 +103,7 @@ exports.UnsafeEmbed = require('@discordjs/builders').UnsafeEmbed;
 exports.Emoji = require('./structures/Emoji').Emoji;
 exports.Guild = require('./structures/Guild').Guild;
 exports.GuildAuditLogs = require('./structures/GuildAuditLogs');
-exports.GuildAuditLogsEntry = require('./structures/GuildAuditLogs').Entry;
+exports.GuildAuditLogsEntry = require('./structures/GuildAuditLogsEntry');
 exports.GuildBan = require('./structures/GuildBan');
 exports.GuildChannel = require('./structures/GuildChannel');
 exports.GuildEmoji = require('./structures/GuildEmoji');
